Changed ObjectSpaceship upgrade interface to use the Upgrade class.

+++ src/WidgetWeaponStatus.cc
@@ -7,13 +7,13 @@ WidgetWeaponStatus::WidgetWeaponStatus(Widget * parent, const glm::ivec2 & posit
   : WidgetSpaceshipStatus(parent, position, size)
 {
   delete _meter;
-  _meter = new MeshProgressBar(GetSize().x, GetSize().y, glm::vec3(1, 0, 0));
+  _meter = new MeshProgressBar(static_cast<float>(GetSize().x), static_cast<float>(GetSize().y), glm::vec3(1, 0, 0));
 }
 
 
 void WidgetWeaponStatus::Draw() const
 {
-  _meter->SetValue(std::clamp(_spaceship->GetWeaponHeat(0), 0.0, 1.0));
+  _meter->SetValue(static_cast<float>(std::clamp(_spaceship->GetWeaponHeat(0), 0.0, 1.0)));
   _meter->Draw(GetModel(), GetView(), GetProjection(), GetMVP());
   
   Widget::Draw();
